security created

+++ app/controllers/drafts_controller.rb
@@ -1,43 +1,61 @@
 class DraftsController < ApplicationController
-
   def new
+    if session[:admin_id]
       @draft = Draft.new
       render :layout => 'admins'
+    else
+      redirect_to '/admin/login'
+    end
   end
 
   def edit
-    @draft = Draft.find_by(id: params[:id])
-    render :layout => 'admins'
+    if session[:admin_id]
+      @draft = Draft.find_by(id: params[:id])
+      render :layout => 'admins'
+    else
+      redirect_to '/admin/login'
+    end
   end
 
   def destroy
-    draft = Draft.find_by(id: params[:id])
-    draft.destroy
-    redirect_to '/admin'
-  end
-
-  def update
-    
-    if params[:commit] == "SAVE"
-      draft = Draft.find_by(id: params[:id])
-      draft.update(draft_params)
-      draft.save
-      redirect_to '/admin'
-    elsif params[:commit] == "PUBLISH"
-      @post = Post.create(draft_params)
+    if session[:admin_id]
       draft = Draft.find_by(id: params[:id])
       draft.destroy
       redirect_to '/admin'
+    else
+      redirect_to '/admin/login'
+    end
+  end
+
+  def update
+    if session[:admin_id]
+      if params[:commit] == "SAVE"
+        draft = Draft.find_by(id: params[:id])
+        draft.update(draft_params)
+        draft.save
+        redirect_to '/admin'
+      elsif params[:commit] == "PUBLISH"
+        @post = Post.create(draft_params)
+        draft = Draft.find_by(id: params[:id])
+        draft.destroy
+        redirect_to '/admin'
+      end
+    else
+      redirect_to '/admin/login'
     end
   end
 
   def create
-    if params[:commit] == "SAVE"
-      @draft = Draft.create(draft_params)
-      redirect_to '/admin'
-    elsif params[:commit] == "PUBLISH"
-      @post = Post.create(draft_params)
-      redirect_to '/admin'
+    if session[:admin_id]
+      if params[:commit] == "SAVE"
+        @draft = Draft.create(draft_params)
+        redirect_to '/admin'
+      elsif params[:commit] == "PUBLISH"
+        @post = Post.create(draft_params)
+        redirect_to '/admin'
+      end
+    else
+      redirect_to '/admin/login'
     end
   end
